requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.1.12
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
djangorestframework==3.12.4
easy-mask==1.3.4
idna==2.10
Pillow==8.2.0
pkg-resources==0.0.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
